[Refactor] 분배 메서드 admin에서 분배 페이지로 이동
