+++ .env
@@ -1,12 +1,12 @@
     
-DB_HOST = cbl.csnv6qgbbdm9.ap-northeast-2.rds.amazonaws.com
+DB_HOST = cbl.cdjpvvwi26cz.ap-northeast-2.rds.amazonaws.com
 DB_PORT = 5432
 DB_USERNAME = postgres
 DB_PASSWORD = creatorbear1
 DB_DATABASE = postgres
 
 
-AWS_S3_BUCKET_NAME= depthes
-AWS_ACCESS_KEY_ID= AKIAXC6KU5WE65L4KRMY
-AWS_SECRET_ACCESS_KEY= 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
+AWS_S3_BUCKET_NAME= cblprofile
+AWS_ACCESS_KEY_ID= AKIA3E67MKOAJL34JDGI
+AWS_SECRET_ACCESS_KEY= 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
 AWS_REGION= ap-northeast-2
