+++ src/admin/distribution/distribution.controller.ts
@@ -1,18 +1,16 @@
 import { Controller, Get, Param } from '@nestjs/common';
 import { ApiCreatedResponse, ApiOperation } from '@nestjs/swagger';
+import { Category } from '../entities/category.entity';
 import { DistributionService } from './distribution.service';
 
 @Controller('distribution')
 export class DistributionController {
     constructor(private readonly distributionService: DistributionService) { }
-
-    // // 카테고리 또는 작가 별 작품 리스트 가져오기
-    // @Get(':id')
-    // @ApiOperation({ summary: '카테고리 또는 작가 별 작품 리스트 가져오는 API', description: '카테고리 또는 작가 별 작품 리스트 가져오기' })
-    // @ApiCreatedResponse({ description: '카테고리 또는 작가 별 작품 리스트 가져오기' })
-
-    // calculateDistribution(@Param('id') id: string): Promise<void> {
-        
-    // }
-
+    // 작가 별 분배포인트 계산
+    @Get('/:id')
+    @ApiOperation({ summary: '작가 별 분배포인트 계산 API', description: '작가 별 분배포인트 계산' })
+    @ApiCreatedResponse({ description: '작가 별 분배포인트 계산', type: Category })
+    calculateDistribution(@Param('id') id: number): Promise<number> {
+        return this.distributionService.calculateDistribution(id);
+    }
 }
